test(quiz): behaviour-first titles, split attempts tests, typed router mock
Three test-hygiene issues batched into one PR — all test-only, plus one
new test-support module. No production source is touched.

#1155 — seven `it(...)` titles named storage mechanics (sessionStorage,
the handoff payload, the user-scoped key) rather than observable
behaviour, against code-style.md §7. Renamed to describe outcomes, e.g.
"writes session data to sessionStorage after a successful start" becomes
"carries the started session into the runner". No assertion changed.

Two details worth recording. The `study-start-handlers` title claimed
"before navigating", but that test asserts no navigation — `router.push`
is asserted by the preceding sibling test — so the new title drops the
claim rather than restating it. And the renamed
`quiz-recovery-handlers` test carried a `// Key must be scoped to the
user` comment that merely paraphrased the old title; §7 "Test Comments:
Audit After Renaming" requires dropping those, so it is deleted.

The `does not call <X>` titles are deliberately untouched — §7 lists
them as PERMITTED contracts, and renaming them has been a past mistake.

#1154 — `internal-exams/queries.test.ts` tested TWO source modules,
against §7 "One Test File Per Source File"; `attempts-queries.ts` had no
co-located suite at all. The `listInternalExamAttempts` block moves to a
new `attempts-queries.test.ts`. Shared scaffolding is DUPLICATED rather
than moved, so each file stands alone. Assertions are byte-identical to
what was removed; 17 + 15 = 32 cases, unchanged.

#1157 — most router mocks bypass structural type-checking, so a breaking
change to Next's `AppRouterInstance` does not surface. When Next 16.3
added a required `bfcacheId`, only 2 of 44 mock sites failed to compile.

Adds `apps/web/lib/test-support/mock-router.ts` exporting
`createMockRouter(overrides): AppRouterInstance`. Its explicit return-type
annotation on a direct object literal is the point: a future required
member fails `check-types` in exactly one place. Ships with a co-located
test, per §7's rule that new `lib/` utilities carry tests in the same
commit.

Converts the 7 sites that can use it — the 5 `as unknown as` double-casts
and the 2 that hand-rolled the full member literal.

The 37 `vi.mock('next/navigation')` factory sites are NOT converted.
Vitest hoists `vi.mock` factories above imports, so they cannot reference
an imported helper without restructuring each through `vi.hoisted` — a
per-file refactor, not a substitution. Issue #1157's AC #4 explicitly
permits leaving them with a rationale, and the single typed factory
already satisfies AC #2. Tracked in #1159.

Verification: 70 test files / 1125 tests pass; `check-types` clean;
`biome check` clean across all 15 files; the test-title-leakage hook
reports no new violating titles.

Closes #1154
Closes #1155
Closes #1157
